created firstActivity for checking if device has camera. created secondActivity for building camera app.
