+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".FirstActivity" />
+        <activity android:name=".SecondActivity"></activity>
     </application>
 
 </manifest>
