+++ app/src/main/java/com/holmeswareapps/friendfinder/SecondActivity.java
@@ -1,0 +1,14 @@
+package com.holmeswareapps.friendfinder;
+
+import android.app.Activity;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class SecondActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.second_activity);
+    }
+}
