+++ app/src/main/res/layout/first_activity.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    tools:context="com.holmeswareapps.friendfinder.FirstActivity">
+
+    <TextView
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="10dp"
+        android:paddingLeft="10dp"
+        android:lines="2"
+        android:id="@+id/hasCamera"
+        android:clickable="true"
+        android:autoLink="all"
+        android:fontFamily="Arial"
+        android:typeface="serif"
+        android:textSize="20sp">
+    </TextView>
+
+</RelativeLayout>
